channellist.h: fix type conversion
Wconversion

+++ src/gui/channellist.h
@@ -257,7 +257,7 @@ public:
 		SORT_CH_NUMBER,
 		SORT_MAX
 	};
-	unsigned Size() { return (*chanlist).size(); }
+	size_t Size() { return (*chanlist).size(); }
 	ZapitChannelList &getChannels() { return channels; };
 	bool checkLockStatus(neutrino_msg_data_t data, bool pip = false);
 	CComponentsHeader* getHeaderObject();
